The main python 3 script.  Requires serial and time.
Hopefully someone will find this code helpful.  

This code is used to take the HPGL output of inkcut and send it over a serial (COM) port to a Gerber GS750. (It will probably work for other models)  This script iterates through the file and attempts to keep from overwhelming the buffer of the plotter by waiting a little bit between sending each command.  Be sure to set the settings in the script to match your set-up.  If there are any improvements or optimizations, I'd love to see them.